Use nebulex instead cache adapters
