+++ lib/gamora/cache/introspect.ex
@@ -4,7 +4,7 @@ defmodule Gamora.Cache.Introspect do
   data in the cache in order to improve performance.
   """
 
-  use Gamora.Cache
+  use Gamora.Cache.Actions
 
   alias Gamora.API
   alias Gamora.Plugs.AuthenticatedUser
@@ -28,6 +28,6 @@ defmodule Gamora.Cache.Introspect do
 
   defp expires_in do
     Application.get_env(:ueberauth, AuthenticatedUser, [])
-    |> Keyword.get(:introspect_cache_expires_in, 0)
+    |> Keyword.get(:introspect_cache_expires_in, :timer.seconds(0))
   end
 end
